refactor: cart 모델 함수들 분리
